+++ clients/spot/src/main/java/com/binance/connector/client/spot/rest/api/UserDataStreamApi.java
@@ -44,7 +44,7 @@ public class UserDataStreamApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-spot/1.2.0 (Java/%s; %s; %s)",
+                    "binance-spot/2.0.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = true;