+++ clients/spot/src/main/java/com/binance/connector/client/spot/websocket/api/model/MyTradesResponse.java
@@ -65,7 +65,7 @@ public class MyTradesResponse extends BaseDTO {
 
     @SerializedName(SERIALIZED_NAME_RATE_LIMITS)
     @jakarta.annotation.Nullable
-    private List<@Valid ExchangeInfoResponseRateLimitsInner> rateLimits;
+    private List<@Valid AccountCommissionResponseRateLimitsInner> rateLimits;
 
     public MyTradesResponse() {}
 
@@ -139,12 +139,13 @@ public class MyTradesResponse extends BaseDTO {
 
     public MyTradesResponse rateLimits(
             @jakarta.annotation.Nullable
-                    List<@Valid ExchangeInfoResponseRateLimitsInner> rateLimits) {
+                    List<@Valid AccountCommissionResponseRateLimitsInner> rateLimits) {
         this.rateLimits = rateLimits;
         return this;
     }
 
-    public MyTradesResponse addRateLimitsItem(ExchangeInfoResponseRateLimitsInner rateLimitsItem) {
+    public MyTradesResponse addRateLimitsItem(
+            AccountCommissionResponseRateLimitsInner rateLimitsItem) {
         if (this.rateLimits == null) {
             this.rateLimits = new ArrayList<>();
         }
@@ -159,13 +160,13 @@ public class MyTradesResponse extends BaseDTO {
      */
     @jakarta.annotation.Nullable
     @Valid
-    public List<@Valid ExchangeInfoResponseRateLimitsInner> getRateLimits() {
+    public List<@Valid AccountCommissionResponseRateLimitsInner> getRateLimits() {
         return rateLimits;
     }
 
     public void setRateLimits(
             @jakarta.annotation.Nullable
-                    List<@Valid ExchangeInfoResponseRateLimitsInner> rateLimits) {
+                    List<@Valid AccountCommissionResponseRateLimitsInner> rateLimits) {
         this.rateLimits = rateLimits;
     }
 
@@ -220,7 +221,7 @@ public class MyTradesResponse extends BaseDTO {
             String resultValueAsString = JSON.getGson().toJson(resultValue);
             valMap.put("result", resultValueAsString);
         }
-        List<@Valid ExchangeInfoResponseRateLimitsInner> rateLimitsValue = getRateLimits();
+        List<@Valid AccountCommissionResponseRateLimitsInner> rateLimitsValue = getRateLimits();
         if (rateLimitsValue != null) {
             String rateLimitsValueAsString = JSON.getGson().toJson(rateLimitsValue);
             valMap.put("rateLimits", rateLimitsValueAsString);
@@ -358,7 +359,7 @@ public class MyTradesResponse extends BaseDTO {
 
                 // validate the optional field `rateLimits` (array)
                 for (int i = 0; i < jsonArrayrateLimits.size(); i++) {
-                    ExchangeInfoResponseRateLimitsInner.validateJsonElement(
+                    AccountCommissionResponseRateLimitsInner.validateJsonElement(
                             jsonArrayrateLimits.get(i));
                 }
                 ;