+++ clients/spot/src/main/java/com/binance/connector/client/spot/websocket/api/model/OrderListPlaceOtocoResponseResultOrderReportsInner.java
@@ -136,6 +136,12 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
     @jakarta.annotation.Nullable
     private String selfTradePreventionMode;
 
+    public static final String SERIALIZED_NAME_STOP_PRICE = "stopPrice";
+
+    @SerializedName(SERIALIZED_NAME_STOP_PRICE)
+    @jakarta.annotation.Nullable
+    private String stopPrice;
+
     public OrderListPlaceOtocoResponseResultOrderReportsInner() {}
 
     public OrderListPlaceOtocoResponseResultOrderReportsInner symbol(
@@ -459,6 +465,26 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
         this.selfTradePreventionMode = selfTradePreventionMode;
     }
 
+    public OrderListPlaceOtocoResponseResultOrderReportsInner stopPrice(
+            @jakarta.annotation.Nullable String stopPrice) {
+        this.stopPrice = stopPrice;
+        return this;
+    }
+
+    /**
+     * Get stopPrice
+     *
+     * @return stopPrice
+     */
+    @jakarta.annotation.Nullable
+    public String getStopPrice() {
+        return stopPrice;
+    }
+
+    public void setStopPrice(@jakarta.annotation.Nullable String stopPrice) {
+        this.stopPrice = stopPrice;
+    }
+
     @Override
     public boolean equals(Object o) {
         if (this == o) {
@@ -510,7 +536,10 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
                         orderListPlaceOtocoResponseResultOrderReportsInner.workingTime)
                 && Objects.equals(
                         this.selfTradePreventionMode,
-                        orderListPlaceOtocoResponseResultOrderReportsInner.selfTradePreventionMode);
+                        orderListPlaceOtocoResponseResultOrderReportsInner.selfTradePreventionMode)
+                && Objects.equals(
+                        this.stopPrice,
+                        orderListPlaceOtocoResponseResultOrderReportsInner.stopPrice);
     }
 
     @Override
@@ -531,7 +560,8 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
                 type,
                 side,
                 workingTime,
-                selfTradePreventionMode);
+                selfTradePreventionMode,
+                stopPrice);
     }
 
     @Override
@@ -558,6 +588,7 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
         sb.append("		selfTradePreventionMode: ")
                 .append(toIndentedString(selfTradePreventionMode))
                 .append("\n");
+        sb.append("		stopPrice: ").append(toIndentedString(stopPrice)).append("\n");
         sb.append("}");
         return sb.toString();
     }
@@ -646,6 +677,11 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
             String selfTradePreventionModeValueAsString = selfTradePreventionModeValue.toString();
             valMap.put("selfTradePreventionMode", selfTradePreventionModeValueAsString);
         }
+        String stopPriceValue = getStopPrice();
+        if (stopPriceValue != null) {
+            String stopPriceValueAsString = stopPriceValue.toString();
+            valMap.put("stopPrice", stopPriceValueAsString);
+        }
 
         valMap.put("timestamp", getTimestamp());
         return asciiEncode(
@@ -721,6 +757,10 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
         if (selfTradePreventionModeValue != null) {
             valMap.put("selfTradePreventionMode", selfTradePreventionModeValue);
         }
+        Object stopPriceValue = getStopPrice();
+        if (stopPriceValue != null) {
+            valMap.put("stopPrice", stopPriceValue);
+        }
 
         valMap.put("timestamp", getTimestamp());
         return valMap;
@@ -763,6 +803,7 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
         openapiFields.add("side");
         openapiFields.add("workingTime");
         openapiFields.add("selfTradePreventionMode");
+        openapiFields.add("stopPrice");
 
         // a set of required properties/fields (JSON key names)
         openapiRequiredFields = new HashSet<String>();
@@ -902,6 +943,14 @@ public class OrderListPlaceOtocoResponseResultOrderReportsInner extends BaseDTO 
                             "Expected the field `selfTradePreventionMode` to be a primitive type in"
                                     + " the JSON string but got `%s`",
                             jsonObj.get("selfTradePreventionMode").toString()));
+        }
+        if ((jsonObj.get("stopPrice") != null && !jsonObj.get("stopPrice").isJsonNull())
+                && !jsonObj.get("stopPrice").isJsonPrimitive()) {
+            throw new IllegalArgumentException(
+                    String.format(
+                            "Expected the field `stopPrice` to be a primitive type in the JSON"
+                                    + " string but got `%s`",
+                            jsonObj.get("stopPrice").toString()));
         }
     }