+++ examples/spot/src/main/java/com/binance/connector/client/spot/rest/trade/OrderAmendKeepPriorityExample.java
@@ -40,8 +40,9 @@ public class OrderAmendKeepPriorityExample {
     /**
      * Order Amend Keep Priority
      *
-     * <p>Reduce the quantity of an existing open order. Read [Order Amend Keep Priority
-     * FAQ](faqs/order_amend_keep_priority.md) to learn more. Weight: 1
+     * <p>Reduce the quantity of an existing open order. This adds 0 orders to the
+     * &#x60;EXCHANGE_MAX_ORDERS&#x60; filter and the &#x60;MAX_NUM_ORDERS&#x60; filter. Read [Order
+     * Amend Keep Priority FAQ](faqs/order_amend_keep_priority.md) to learn more. Weight: 4
      *
      * @throws ApiException if the Api call fails
      */