+++ examples/spot/src/main/java/com/binance/connector/client/spot/rest/trade/OrderListOtocoExample.java
@@ -51,9 +51,9 @@ public class OrderListOtocoExample {
      * has 2 pending orders (pending above and pending below), forming an OCO pair. The pending
      * orders are only placed on the order book when the working order gets **fully filled**. * The
      * rules of the pending above and pending below follow the same rules as the [Order list
-     * OCO](#new-order-list---oco-trade). * OTOCOs add **3 orders** against the unfilled order
-     * count, &#x60;EXCHANGE_MAX_NUM_ORDERS&#x60; filter, and &#x60;MAX_NUM_ORDERS&#x60; filter.
-     * Weight: 1
+     * OCO](#new-order-list---oco-trade). * OTOCOs add **3 orders** to the
+     * &#x60;EXCHANGE_MAX_NUM_ORDERS&#x60; filter and &#x60;MAX_NUM_ORDERS&#x60; filter. Weight: 1
+     * Unfilled Order Count: 3
      *
      * @throws ApiException if the Api call fails
      */