+++ examples/spot/src/main/java/com/binance/connector/client/spot/rest/trade/SorOrderExample.java
@@ -42,8 +42,9 @@ public class SorOrderExample {
     /**
      * New order using SOR
      *
-     * <p>Places an order using smart order routing (SOR). Read [SOR FAQ](faqs/sor_faq.md) to learn
-     * more. Weight: 1
+     * <p>Places an order using smart order routing (SOR). This adds 1 order to the
+     * &#x60;EXCHANGE_MAX_ORDERS&#x60; filter and the &#x60;MAX_NUM_ORDERS&#x60; filter. Read [SOR
+     * FAQ](faqs/sor_faq.md) to learn more. Weight: 1 Unfilled Order Count: 1
      *
      * @throws ApiException if the Api call fails
      */