+++ examples/spot/src/main/java/com/binance/connector/client/spot/websocket/api/trade/OrderListPlaceOcoExample.java
@@ -55,8 +55,8 @@ public class OrderListPlaceOcoExample {
      * If the OCO is on the &#x60;BUY&#x60; side: * &#x60;LIMIT_MAKER&#x60; &#x60;price&#x60; &lt;
      * Last Traded Price &lt; &#x60;STOP_LOSS/STOP_LOSS_LIMIT&#x60; &#x60;stopPrice&#x60; *
      * &#x60;TAKE_PROFIT stopPrice&#x60; &gt; Last Traded Price &gt; &#x60;STOP_LOSS/STOP_LOSS_LIMIT
-     * stopPrice&#x60; * OCOs add **2 orders** to the unfilled order count,
-     * &#x60;EXCHANGE_MAX_ORDERS&#x60; filter, and &#x60;MAX_NUM_ORDERS&#x60; filter. Weight: 1
+     * stopPrice&#x60; * OCOs add **2 orders** to the &#x60;EXCHANGE_MAX_ORDERS&#x60; filter and
+     * &#x60;MAX_NUM_ORDERS&#x60; filter. Weight: 1 Unfilled Order Count: 2
      */
     public void orderListPlaceOcoExampleAsync() {
         OrderListPlaceOcoRequest orderListPlaceOcoRequest = new OrderListPlaceOcoRequest();
@@ -91,8 +91,8 @@ public class OrderListPlaceOcoExample {
      * If the OCO is on the &#x60;BUY&#x60; side: * &#x60;LIMIT_MAKER&#x60; &#x60;price&#x60; &lt;
      * Last Traded Price &lt; &#x60;STOP_LOSS/STOP_LOSS_LIMIT&#x60; &#x60;stopPrice&#x60; *
      * &#x60;TAKE_PROFIT stopPrice&#x60; &gt; Last Traded Price &gt; &#x60;STOP_LOSS/STOP_LOSS_LIMIT
-     * stopPrice&#x60; * OCOs add **2 orders** to the unfilled order count,
-     * &#x60;EXCHANGE_MAX_ORDERS&#x60; filter, and &#x60;MAX_NUM_ORDERS&#x60; filter. Weight: 1
+     * stopPrice&#x60; * OCOs add **2 orders** to the &#x60;EXCHANGE_MAX_ORDERS&#x60; filter and
+     * &#x60;MAX_NUM_ORDERS&#x60; filter. Weight: 1 Unfilled Order Count: 2
      */
     public void orderListPlaceOcoExampleSync() {
         OrderListPlaceOcoRequest orderListPlaceOcoRequest = new OrderListPlaceOcoRequest();