+++ clients/spot/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.0.0 - 2025-05-14
+
+- Updated response types.
+- Updated request parameters to correctly specify their required status.
+
 ## 1.2.0 - 2025-05-13
 
 - Add proxy authentication for websocket